Move classes for reports page into separate file
They were in the routes while I was extracting but they can't stay there
long term, so moving them to their own file.
